Certificates revamped
