+++ src/providers/sh/commands/alias/get-app-last-deployment.js
@@ -1,11 +1,11 @@
 // @flow
 import chalk from 'chalk'
 import wait from '../../../../util/output/wait'
+import { Output, Now } from '../../util/types'
+import type { User } from '../../util/types'
 
-import { Output, Now } from './types'
 import fetchDeploymentByIdOrHost from './get-deployment-by-id-or-host'
 import fetchDeploymentsByAppName from './get-deployments-by-appname'
-import type { User } from './types'
 
 async function getAppLastDeployment(output: Output, now: Now, appName: string, user: User, contextName: string) {
   output.debug(`Looking for deployments matching app ${appName}`)
