+++ src/providers/sh/commands/alias/get-app-name.js
@@ -1,7 +1,7 @@
 // @flow
 import path from 'path'
-import { NowError } from './now-error'
-import { Output } from './types'
+import { NowError } from '../../util/now-error'
+import { Output } from '../../util/types'
 import getConfig from './get-config'
 import readPackage from './read-package'
