+++ src/providers/sh/commands/alias/get-domain-status.js
@@ -1,6 +1,6 @@
 // @flow
 import qs from 'querystring'
-import { Now } from './types'
+import { Now } from '../../util/types'
 
 type DomainStatus = {
   available: boolean,
