+++ src/providers/sh/commands/certs/add.js
@@ -1,0 +1,104 @@
+// @flow
+import chalk from 'chalk'
+
+import Now from '../../util'
+import getContextName from '../../util/get-context-name'
+import stamp from '../../../../util/output/stamp'
+import wait from '../../../../util/output/wait'
+import { CLIContext, Output } from '../../util/types'
+import * as Errors from '../../util/errors'
+import { handleDomainConfigurationError } from '../../util/error-handlers'
+import type { CLICertsOptions } from '../../util/types'
+
+import createCertFromFile from '../../util/certs/create-cert-from-file'
+import createCertForCns from '../../util/certs/create-cert-for-cns'
+
+async function add(ctx: CLIContext, opts: CLICertsOptions, args: string[], output: Output): Promise<number> {
+  const {authConfig: { credentials }, config: { sh }} = ctx
+  const { currentTeam } = sh;
+  const { apiUrl } = ctx;
+  const contextName = getContextName(sh);
+  const addStamp = stamp()
+  let cert
+  
+  const {
+    ['--overwrite']: overwite,
+    ['--debug']: debugEnabled,
+    ['--crt']: crtPath,
+    ['--key']: keyPath,
+    ['--ca']: caPath,
+  } = opts;
+  
+  // $FlowFixMe
+  const {token} = credentials.find(item => item.provider === 'sh')
+  const now = new Now({ apiUrl, token, debug: debugEnabled, currentTeam })
+  
+  if (overwite) {
+    output.error('Overwrite option is deprecated')
+    now.close();
+    return 1;
+  }
+
+  if (crtPath || keyPath || caPath) {
+    if ((args.length !== 0) || (!crtPath || !keyPath || !caPath)) {
+      output.error(`Invalid number of arguments to create a custom certificate entry. Usage:`)
+      output.print(`  ${chalk.cyan(`now certs add --crt <domain.crt> --key <domain.key> --ca <ca.crt>`)}\n`)
+      now.close();
+      return 1
+    }
+    
+    // Create a custom certificate from the given file paths
+    cert = await createCertFromFile(now, keyPath, crtPath, caPath, contextName)
+    if (cert instanceof Errors.InvalidCert) {
+      output.error(`The provided certificate is not valid and can't be added.`)
+      return 1
+    } else if (cert instanceof Errors.DomainPermissionDenied) {
+      output.error(`You don't have permissions over domain ${chalk.underline(cert.meta.domain)} under ${chalk.bold(cert.meta.context)}.`)
+      return 1
+    }
+  } else {
+    if (args.length < 1) {
+      output.error(`Invalid number of arguments to create a custom certificate entry. Usage:`)
+      output.print(`  ${chalk.cyan(`now certs add <cn>[, <cn>]`)}\n`)
+      now.close();
+      return 1
+    }
+
+    // Create the certificate from the given array of CNs
+    const cancelWait = wait(`Issuing a certificate for ${chalk.bold(args.join(', '))}`);
+    cert = await createCertForCns(now, args, contextName)
+    cancelWait();
+    if (cert instanceof Errors.TooManyRequests) {
+      output.error(`Too many requests detected for ${cert.meta.api} API. Try again later.`)
+      return 1
+    } else if (cert instanceof Errors.TooManyCertificates) {
+      output.error(`Too many certificates already issued for exact set of domains: ${cert.meta.domains.join(', ')}`)
+      return 1
+    } else if (cert instanceof Errors.DomainValidationRunning) {
+      output.error(`There is a validation in course for ${chalk.underline(cert.meta.domain)}. Wait until it finishes.`)
+      return 1
+    } else if (cert instanceof Errors.DomainConfigurationError) {
+      handleDomainConfigurationError(output, cert)
+      return 1
+    } else if (cert instanceof Errors.CantGenerateWildcardCert) {
+      output.error(`Wildcard certificates are allowed only for domains in ${chalk.underline('zeit.world')}`)
+      return 1
+    } else if (cert instanceof Errors.DomainsShouldShareRoot) {
+      output.error(`All given common names should share the same root domain.`)
+      return 1
+    } else if (cert instanceof Errors.InvalidWildcardDomain) {
+      output.error(`Invalid domain ${chalk.underline(cert.meta.domain)}. Wildcard domains can only be followed by a root domain.`)
+      return 1
+    } else if (cert instanceof Errors.DomainPermissionDenied) {
+      output.error(`You don't have permissions over domain ${chalk.underline(cert.meta.domain)} under ${chalk.bold(cert.meta.context)}.`)
+      return 1
+    }
+  }
+
+  // Print success message
+  const cns = chalk.bold(cert.cns.join(', '))
+  output.success(`Certificate entry for ${cns} created ${addStamp()}`)
+  return 0
+}
+
+export default add
