+++ src/providers/sh/util/arg-common.js
@@ -1,25 +1,25 @@
-// common options supported by all subcommands
+// @flow
 
-module.exports = () => (
-  {
-    '--help': Boolean,
-    '-h': '--help',
+const ARG_COMMON = {
+  '--help': Boolean,
+  '-h': '--help',
 
-    '--debug': Boolean,
-    '-d': '--debug',
+  '--debug': Boolean,
+  '-d': '--debug',
 
-    '--token': String,
-    '-t': '--token',
+  '--token': String,
+  '-t': '--token',
 
-    '--team': String,
-    '-T': '--team',
+  '--team': String,
+  '-T': '--team',
 
-    '--local-config': String,
-    '-A': '--local-config',
+  '--local-config': String,
+  '-A': '--local-config',
 
-    '--global-config': String,
-    '-Q': '--global-config',
+  '--global-config': String,
+  '-Q': '--global-config',
 
-    '--api': String
-  }
-)
+  '--api': String
+}
+
+module.exports = () => ARG_COMMON
