+++ src/providers/sh/util/error-handlers.js
@@ -1,0 +1,33 @@
+// @flow
+import chalk from 'chalk'
+import table from 'text-table'
+
+import strlen from './strlen'
+import { Output } from './types'
+import { DomainConfigurationError } from './errors'
+
+export function handleDomainConfigurationError(output: Output, error: DomainConfigurationError) {
+  output.error(`We couldn't verify the propagation of the DNS settings for ${chalk.underline(error.meta.domain)}`)
+  if (error.meta.external) {
+    output.print(`  The propagation may take a few minutes, but please verify your settings:\n\n`)
+    output.print(dnsTable([
+      error.meta.subdomain === null
+        ? ['', 'ALIAS', 'alias.zeit.co']
+        : [error.meta.subdomain, 'CNAME', 'alias.zeit.co']  
+    ]) + '\n');
+  } else {
+    output.print(`  We configured them for you, but the propagation may take a few minutes.\n`)
+    output.print(`  Please try again later.\n`)
+  }
+}
+
+function dnsTable(rows, extraSpace = '') {
+  return table([
+    ['name', 'type', 'value'].map(v => chalk.gray(v)),
+    ...rows
+  ], {
+    align: ['l', 'l', 'l'],
+    hsep: ' '.repeat(8),
+    stringLength: strlen
+  }).replace(/^(.*)/gm, `${extraSpace}  $1`)
+}
