+++ src/util/output/elapsed.js
@@ -1,7 +1,8 @@
-const ms = require('ms')
-const chalk = require('chalk')
+// @flow
+import ms from 'ms'
+import chalk from 'chalk'
 
 // styles the "[30ms]" string based on a number of ms
-module.exports = function elapsed (time) {
+module.exports = function elapsed (time: number): string {
   return chalk.gray(`[${time < 1000 ? `${time}ms` : ms(time)}]`);
 }
